Consolidate channel histogram component
