--- app-frontend/src/app/components/histogram/channelHistogram/channelHistogram.component.js
@@ -1,13 +1,0 @@
-import channelHistogramTpl from './channelHistogram.html';
-
-const channelHistogram = {
-    templateUrl: channelHistogramTpl,
-    controller: 'ChannelHistogramController',
-    bindings: {
-        data: '<',
-        corrections: '<',
-        onChange: '&'
-    }
-};
-
-export default channelHistogram;
